+++ libs/cardano-protocol/CHANGELOG.md
@@ -16,4 +16,4 @@
 
 ### `testlib`
 
-* Add `testlib` with `Test.Cardano.Protocol.Praos.Arbitrary` and `Test.Cardano.Protocol.Leios.Arbitrary`, providing `Arbitrary` instances for the Praos and Leios `Header`/`HeaderBody`, `EbAnnouncement`, `InputVRF`, `OCert`, `KESPeriod` and `PrevHash`, and non-annotator `DecCBOR` instances for the Praos and Leios `Header`
+* Add `testlib` with `Test.Cardano.Protocol.Praos.Arbitrary` and `Test.Cardano.Protocol.Leios.Arbitrary`, providing `Arbitrary` instances for the Praos and Leios `Header`/`HeaderBody`/`Block`, `EbAnnouncement`, `InputVRF`, `OCert`, `KESPeriod` and `PrevHash`, and non-annotator `DecCBOR` instances for the Praos and Leios `Header`